Separated version of TiltEstimator for humanoids
